requirements.txt
crewai>=0.11.0
langchain>=0.1.0
langchain-community>=0.0.13
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain-openai>=0.0.2